+++ SQLiteStudio3/guiSQLiteStudio/selectabledbobjmodel.cpp
@@ -56,7 +56,6 @@ Qt::ItemFlags SelectableDbObjModel::flags(const QModelIndex& idx) const
         case DbTreeItem::Type::COLUMNS:
         case DbTreeItem::Type::COLUMN:
         case DbTreeItem::Type::ITEM_PROTOTYPE:
-        case DbTreeItem::Type::SIGNATURE_OF_THIS:
             break;
     }
 
@@ -136,7 +135,6 @@ bool SelectableDbObjModel::filterAcceptsRow(int srcRow, const QModelIndex& srcPa
         case DbTreeItem::Type::INDEX:
         case DbTreeItem::Type::TRIGGER:
         case DbTreeItem::Type::ITEM_PROTOTYPE:
-        case DbTreeItem::Type::SIGNATURE_OF_THIS:
             return false;
     }
     return false;